requirements.txt
coverage==6.5.0
deepdiff==5.8.1
flake8==5.0.4
mccabe==0.7.0
ordered-set==4.1.0
pip==22.1.2
pycodestyle==2.9.1
pyflakes==2.5.0
pyperclip==1.8.2
PyYAML==6.0
setuptools==63.1.0
wheel==0.37.1
